Create classes to handle DISORT input
I split Output into Output and Unsure---arrays that I'm not so sure how to handle and therefore stuffed with 0s. I added a size class to take care of the "optional" parameters. I also added Control and BoundaryConditions to handle DISORT's control variables and variables specifically related to its handling of boundaries.